Task_manager_project/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
mysqlclient==2.2.0
django-rest-framework-simplejwt==5.3.0
drf-yasg==1.21.7
python-decouple==3.8